% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diverse genos selection function.R
\name{div.sub}
\alias{div.sub}
\title{Diverse lines subset}
\usage{
div.sub(K, n.sub = 100, n.gens = 400, n.iters = 20, verbose = T)
}
\arguments{
\item{K}{An nxn kinship matrix.}

\item{n.sub}{number of lines to subset.}

\item{n.gens}{Number of generations of shuffling.}

\item{n.iters}{Number of iterations of algorithm to try. It will automatically output the best one.}

\item{verbose}{Print output - logical.}
}
\value{
Returns a list of objects including:
"Best_subset" A vector of names of the selected subset of lines from the best iteration.
"Best_Max_K" The maximum kinship between lines in the selected subset
"Max_kinships" A list of vectors of maximum kinship values at each generation of shuffling for each iteration. Use these to check you are running enough generations.
}
\description{
Select a subset of lines from a larger panel that you have a kinship matrix for. It will find a subset of \emph{n} unrelated lines that have a minimum kinship between any pair.
}
\examples{
library(FunFunctions)
#load the A kinship matrix for 599 wheat lines from the BGLR package
data(Wheat599_Amat)
subset<-div.sub(K = wheat599,n.sub = 100,n.gens = 200,n.iters = 5,verbose = T)
}
